chore(integrations): move aiohttp_jinja2 and aiomysql to internal package
